back/requirements.txt
fastapi
uvicorn
pymongo
pydantic
motor
python-dotenv
twilio
fastapi-cors